[36] feat : 치료선택 컴포넌트 분리
